rfac: move useLogs to separate file
