+++ dashboard/src/main/home/cluster-dashboard/expanded-chart/status/types.ts
@@ -1,0 +1,19 @@
+export type SelectedPodType = {
+  spec: {
+    [key: string]: any;
+    containers: {
+      [key: string]: any;
+      name: string;
+    }[];
+  };
+  metadata: {
+    name: string;
+    namespace: string;
+    labels: {
+      [key: string]: string;
+    };
+  };
+  status: {
+    phase: string;
+  };
+};
